src/backend/audio_processor.py
"""
Módulo para procesar audio y separar pistas utilizando modelos de AI.
"""
import os
import logging
from typing import Dict, List, Optional, Tuple
import numpy as np
import librosa
import soundfile as sf
from pathlib import Path

from ..utils.config import (
    PROCESSED_DIR, 
    SAMPLE_RATE, 
    DEMUCS_MODEL, 
    DEMUCS_SEGMENT, 
    DEMUCS_OVERLAP
)

# Configurar logging
logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(name)s - %(levelname)s - %(message)s')
logger = logging.getLogger(__name__)

class AudioProcessor:
    """
    Clase para manejar el procesamiento de audio y la separación de pistas.
    """
    def __init__(self, output_dir: str = PROCESSED_DIR, model_name: str = DEMUCS_MODEL):
        """
        Inicializa el procesador de audio.
        
        Args:
            output_dir: Directorio donde se guardarán los archivos procesados.
            model_name: Nombre del modelo de Demucs a utilizar.
        """
        self.output_dir = Path(output_dir)
        self.model_name = model_name
        os.makedirs(self.output_dir, exist_ok=True)
        
        # Demucs debe ser importado después de importar torch
        # y puede no estar instalado aún
        try:
            # Preparación para cuando instalemos demucs
            # import torch
            # from demucs.pretrained import get_model
            # from demucs.apply import apply_model
            
            # self.device = "cuda" if torch.cuda.is_available() else "cpu"
            # logger.info(f"Utilizando dispositivo: {self.device}")
            
            # self.model = get_model(self.model_name).to(self.device)
            # logger.info(f"Modelo {self.model_name} cargado correctamente.")
            
            logger.info("Interfaz de AudioProcessor preparada. Demucs será implementado en una fase posterior.")
            self._demucs_available = False
        except ImportError:
            logger.warning("Demucs no está instalado. La separación de pistas no estará disponible.")
            self._demucs_available = False
    
    def load_audio(self, audio_path: str) -> Tuple[np.ndarray, int]:
        """
        Carga un archivo de audio y lo convierte a la tasa de muestreo estándar.
        
        Args:
            audio_path: Ruta al archivo de audio.
            
        Returns:
            Tupla con el array numpy del audio y la tasa de muestreo.
        """
        logger.info(f"Cargando audio: {audio_path}")
        audio, sr = librosa.load(audio_path, sr=SAMPLE_RATE, mono=False)
        
        # Si el audio es mono, convertirlo a estéreo (2 canales)
        if audio.ndim == 1:
            audio = np.stack([audio, audio])
        
        logger.info(f"Audio cargado: {audio.shape}, SR={sr}")
        return audio, sr
    
    def separate_tracks(self, audio_path: str, progress_callback=None) -> Optional[Dict[str, str]]:
        """
        Separa un archivo de audio en pistas individuales.
        
        Args:
            audio_path: Ruta al archivo de audio.
            progress_callback: Función opcional para reportar el progreso.
            
        Returns:
            Diccionario con los nombres de las pistas y las rutas a los archivos separados,
            o None si hay un error.
        """
        if not self._demucs_available:
            logger.warning("Demucs no está disponible. La separación de pistas es simulada.")
            # En la versión de prueba de concepto, simularemos la separación
            return self._simulate_track_separation(audio_path, progress_callback)
        
        try:
            # Cargar audio
            audio, sr = self.load_audio(audio_path)
            
            # Generar nombre de salida basado en el nombre del archivo original
            audio_name = os.path.splitext(os.path.basename(audio_path))[0]
            output_dir = self.output_dir / audio_name
            os.makedirs(output_dir, exist_ok=True)
            
            # Esto será implementado cuando Demucs esté instalado
            # El código sería similar a:
            """
            # Aplicar el modelo de separación
            audio_torch = torch.tensor(audio, dtype=torch.float32).to(self.device)
            
            # Separar en chunks por limitaciones de memoria
            chunk_size = int(DEMUCS_SEGMENT * sr)
            overlap_size = int(DEMUCS_OVERLAP * sr)
            
            # Aquí vendría el código para aplicar el modelo por chunks
            # y procesar el resultado
            
            # Las pistas separadas se guardarían en los archivos correspondientes
            """
            
            # Por ahora, simularemos la separación
            return self._simulate_track_separation(audio_path, progress_callback)
            
        except Exception as e:
            logger.error(f"Error al separar pistas: {e}")
            return None
    
    def _simulate_track_separation(self, audio_path: str, progress_callback=None) -> Dict[str, str]:
        """
        Simula la separación de pistas para la prueba de concepto.
        
        Args:
            audio_path: Ruta al archivo de audio original.
            progress_callback: Función para reportar progreso.
            
        Returns:
            Diccionario con nombres de pistas y rutas simuladas.
        """
        audio, sr = self.load_audio(audio_path)
        
        # Generar nombres de salida
        audio_name = os.path.splitext(os.path.basename(audio_path))[0]
        output_dir = self.output_dir / audio_name
        os.makedirs(output_dir, exist_ok=True)
        
        # Diferentes instrumentos que Demucs puede separar
        instruments = ['vocals', 'drums', 'bass', 'guitar', 'piano', 'other']
        track_paths = {}
        
        # Simular progreso
        total_steps = len(instruments)
        
        for i, instrument in enumerate(instruments):
            # En la versión simulada, simplemente creamos copias del audio original
            # con diferentes niveles para simular diferentes instrumentos
            
            # Factores de volumen artificial para simular diferentes instrumentos
            volume_factors = {
                'vocals': 0.8,
                'drums': 0.6,
                'bass': 0.7,
                'guitar': 0.5,
                'piano': 0.4,
                'other': 0.3
            }
            
            # Crear una copia modificada del audio para simular la pista separada
            track_audio = audio * volume_factors.get(instrument, 0.5)
            
            # Ruta de salida
            output_path = output_dir / f"{instrument}.wav"
            
            # Guardar el audio
            sf.write(str(output_path), track_audio.T, sr)
            
            track_paths[instrument] = str(output_path)
            
            # Actualizar progreso
            if progress_callback:
                progress_callback(((i + 1) / total_steps) * 100)
            
            logger.info(f"Pista simulada guardada: {output_path}")
        
        return track_paths
    
    def mix_tracks(self, tracks: Dict[str, Tuple[str, float]], output_path: str) -> bool:
        """
        Mezcla varias pistas con diferentes niveles de volumen.
        
        Args:
            tracks: Diccionario con nombres de pistas, rutas y niveles (0.0-1.0).
            output_path: Ruta del archivo de salida.
            
        Returns:
            True si la mezcla fue exitosa, False en caso contrario.
        """
        try:
            # Cargar todas las pistas
            loaded_tracks = {}
            sr = None
            length = None
            
            for name, (track_path, level) in tracks.items():
                if os.path.exists(track_path):
                    audio, track_sr = self.load_audio(track_path)
                    
                    if sr is None:
                        sr = track_sr
                    
                    # Ajustar volumen
                    audio = audio * level
                    
                    loaded_tracks[name] = audio
                    
                    # Guardar la longitud de la pista más corta
                    if length is None or audio.shape[1] < length:
                        length = audio.shape[1]
            
            if not loaded_tracks:
                logger.error("No se pudo cargar ninguna pista para mezclar")
                return False
            
            # Mezclar pistas recortadas a la misma longitud
            mix = None
            for name, audio in loaded_tracks.items():
                if mix is None:
                    mix = audio[:, :length]
                else:
                    mix += audio[:, :length]
            
            # Normalizar para evitar clipping
            max_val = np.max(np.abs(mix))
            if max_val > 0:
                mix = mix / max_val * 0.9  # 90% del máximo para evitar distorsión
            
            # Guardar mezcla
            sf.write(output_path, mix.T, sr)
            logger.info(f"Mezcla guardada en: {output_path}")
            
            return True
            
        except Exception as e:
            logger.error(f"Error al mezclar pistas: {e}")
            return False


# Ejemplo de uso
if __name__ == "__main__":
    # Este código solo se ejecuta si el script es ejecutado directamente
    processor = AudioProcessor()
    
    # Para pruebas, se necesitaría un archivo de audio
    test_file = "/ruta/a/archivo/audio.wav"
    
    if os.path.exists(test_file):
        # Separar pistas
        def progress(percent):
            print(f"Progreso: {percent:.1f}%")
        
        tracks = processor.separate_tracks(test_file, progress)
        
        if tracks:
            print("Pistas separadas:")
            for name, path in tracks.items():
                print(f"- {name}: {path}")
            
            # Crear una mezcla personalizada
            mix_tracks = {
                'vocals': (tracks['vocals'], 1.0),     # Voz al 100%
                'drums': (tracks['drums'], 0.8),      # Batería al 80%
                'bass': (tracks['bass'], 0.7),        # Bajo al 70%
                'guitar': (tracks['guitar'], 0.0),    # Guitarra al 0% (silenciada)
                'piano': (tracks['piano'], 0.5),      # Piano al 50%
                'other': (tracks['other'], 0.3),      # Otros al 30%
            }
            
            output_mix = os.path.join(PROCESSED_DIR, "mezcla_personalizada.wav")
            if processor.mix_tracks(mix_tracks, output_mix):
                print(f"Mezcla guardada en: {output_mix}")
    else:
        print(f"El archivo {test_file} no existe.")
